Move client interface to connection package
