+++ raft.go
@@ -19,6 +19,7 @@ package raft
 
 import "sync"
 import "github.com/chewr/raft/persistance"
+import "github.com/chewr/raft/connection"
 
 // import "bytes"
 // import "encoding/gob"
@@ -47,7 +48,7 @@ type RaftNode interface {
 //
 type Raft struct {
 	mu        sync.Mutex
-	peers     []Client
+	peers     []connection.Client
 	persister persistance.Persister
 	me        int // index into peers[]
 
@@ -179,7 +180,7 @@ func (rf *Raft) Kill() {
 // Make() must return quickly, so it should start goroutines
 // for any long-running work.
 //
-func Make(peers []Client, me int,
+func Make(peers []connection.Client, me int,
 	persister persistance.Persister, applyCh chan ApplyMsg) RaftNode {
 	rf := &Raft{}
 	rf.peers = peers
