+++ testutil/labrpc_adapter.go
@@ -2,13 +2,13 @@ package testutil
 
 import "fmt"
 import "github.com/chewr/6.824-2016/labrpc"
-import "github.com/chewr/raft"
+import "github.com/chewr/raft/connection"
 
 type labrpcAdapter struct {
 	e *labrpc.ClientEnd
 }
 
-func NewLabRpcAdapter(e *labrpc.ClientEnd) raft.Client {
+func NewLabRpcAdapter(e *labrpc.ClientEnd) connection.Client {
 	return &labrpcAdapter{
 		e: e,
 	}
